problem-create-service/requirements.txt
booktype
aioredis
httpx
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
alembic